Calculate Genomic Relatedness Matrix (GRM) for honeybees from Identical By Descent genomic data

Level 0 function that returns Genomic Relatedness Matrix (GRM) for honeybees from Identical By Descent genomic data (tracked alleles since the founders) - see references on the background theory.

calcBeeGRMIbd(x)

Arguments

x

matrix of haplotypes/genomes with allele indicators for the founders coded as 1, 2, ... Haplotypes/genome are in rows and sites are in columns; no missing values are allowed (this is not checked!). Row names are essential (formated as ind_genome as returned by AlphaSimR IBD functions) to infer the individual and their ploidy (see examples)!

Value

a list with a matrix of gametic relatedness coefficients (genome) and a matrix of individual relatedness coefficients (indiv)

References

Grossman and Eisen (1989) Inbreeding, coancestry, and covariance between relatives for X-chromosomal loci. The Journal of Heredity, https://doi.org/10.1093/oxfordjournals.jhered.a110812

Fernando and Grossman (1989) Covariance between relatives for X-chromosomal loci in a population in disequilibrium. Theoretical and Applied Genetics, https://doi.org/10.1007/bf00305821

Fernando and Grossman (1990) Genetic evaluation with autosomal and X-chromosomal inheritance. Theoretical and Applied Genetics, https://doi.org/10.1007/bf00224018

Van Arendonk, Tier, and Kinghorn (1994) Use of multiple genetic markers in prediction of breeding values. Genetics, https://doi.org/10.1093/genetics/137.1.319

Hill and Weir (2011) Variation in actual relationship as a consequence of Mendelian sampling and linkage. Genetics Research, https://doi.org/10.1017/s0016672310000480

Examples

founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)

basePop <- createVirginQueens(founderGenomes)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
drones <- createDrones(x = basePop[1], nInd = 1000)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- createMultiColony(basePop[2:3], n = 2)
#> Error in createMultiColony(basePop[2:3], n = 2): object 'basePop' not found
apiary <- cross(x = apiary, fathers = fatherGroups[c(2, 3)])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- buildUp(x = apiary, nWorkers = 6, nDrones = 3)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- addVirginQueens(x = apiary, nInd = 5)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

haploQ <- getQueenIbdHaplo(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
haploF <- getFathersIbdHaplo(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
haploW <- getWorkersIbdHaplo(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
haploD <- getDronesIbdHaplo(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
haploV <- getVirginQueensIbdHaplo(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
SP$pedigree
#>      [,1] [,2] [,3]

haplo <- rbind(haploQ, haploF, haploW, haploD, haploV)
#> Error in rbind(haploQ, haploF, haploW, haploD, haploV): object 'haploQ' not found

GRMs <- calcBeeGRMIbd(x = haplo)
#> Error in calcBeeGRMIbd(x = haplo): object 'haplo' not found

# library("Matrix"); image(as(GRMs$genome, "Matrix"))
# library("Matrix"); image(as(GRMs$indiv, "Matrix"))

GRMs
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

x <- diag(GRMs$genome)
#> Error in diag(GRMs$genome): object 'GRMs' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

x <- GRMs$genome[lower.tri(x = GRMs$genome, diag = FALSE)]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

x <- diag(GRMs$indiv)
#> Error in diag(GRMs$indiv): object 'GRMs' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

x <- GRMs$indiv[lower.tri(x = GRMs$indiv, diag = FALSE)]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

qI <- getQueen(apiary[[1]])@id
#> Error in is(x, class2 = "Colony"): object 'apiary' not found
fI <- sort(getFathers(apiary[[1]])@id)
#> Error in is(x, class2 = "Pop"): object 'apiary' not found
wI <- sort(getWorkers(apiary[[1]])@id)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found
dI <- sort(getDrones(apiary[[1]])@id)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found
vI <- sort(getVirginQueens(apiary[[1]])@id)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found

qG <- c(t(outer(X = qI, Y = 1:2, FUN = paste, sep = "_")))
#> Error in outer(X = qI, Y = 1:2, FUN = paste, sep = "_"): object 'qI' not found
fG <- paste(fI, 1, sep = "_")
#> Error in paste(fI, 1, sep = "_"): object 'fI' not found
wG <- c(t(outer(X = wI, Y = 1:2, FUN = paste, sep = "_")))
#> Error in outer(X = wI, Y = 1:2, FUN = paste, sep = "_"): object 'wI' not found
dG <- paste(dI, 1, sep = "_")
#> Error in paste(dI, 1, sep = "_"): object 'dI' not found
vG <- c(t(outer(X = vI, Y = 1:2, FUN = paste, sep = "_")))
#> Error in outer(X = vI, Y = 1:2, FUN = paste, sep = "_"): object 'vI' not found

# Queen vs others
GRMs$genome[qG, qG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[qI, qI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[fG, qG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[fI, qI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[wG, qG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[wI, qI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[dG, qG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[dI, qI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[vG, qG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[vI, qI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

# Fathers vs others
GRMs$genome[fG, fG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[fI, fI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[wG, fG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[wI, fI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[dG, fG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[dI, fI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[vG, fG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[vI, fI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

# Workers vs others
GRMs$genome[wG, wG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[wI, wI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[dG, wG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[dI, wI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found

GRMs$genome[vG, wG]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found
GRMs$indiv[vI, wI]
#> Error in eval(expr, envir, enclos): object 'GRMs' not found