Calculate Genomic Relatedness Matrix (GRM) for honeybees from Identical By State genomic data

Level 0 function that returns Genomic Relatedness Matrix (GRM) for honeybees from Identical By State genomic data (bi-allelic SNP represented as allele dosages) following the method for the sex X chromosome (Druet and Legarra, 2020)

calcBeeGRMIbs(x, sex, alleleFreq = NULL)

calcBeeAlleleFreq(x, sex)

Arguments

x

matrix of genotypes represented as allele dosage coded as 0, 1, or 2 in females (queens or workers) and as 0 or 1 in males (fathers or drones); individuals are in rows and sites are in columns; no missing values are allowed (this is not checked - you will get NAs!)

sex

character vector denoting sex for individuals with genotypes in x - "F" for female and "M" for male

alleleFreq

numeric, vector of allele frequencies for the sites in x; if NULL, then calcBeeAlleleFreq is used

Value

matrix of genomic relatedness coefficients

Functions

calcBeeAlleleFreq(): Calculate allele frequencies from honeybee genotypes

References

Druet and Legarra (2020) Theoretical and empirical comparisons of expected and realized relationships for the X-chromosome. Genetics Selection Evolution, 52:50 https://doi.org/10.1186/s12711-020-00570-6

Examples

founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)

basePop <- createVirginQueens(founderGenomes)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
drones <- createDrones(x = basePop[1], nInd = 1000)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- createMultiColony(basePop[2:3], n = 2)
#> Error in createMultiColony(basePop[2:3], n = 2): object 'basePop' not found
apiary <- cross(x = apiary, fathers = fatherGroups[c(2, 3)])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- buildUp(x = apiary, nWorkers = 6, nDrones = 3)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- addVirginQueens(x = apiary, nInd = 5)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

genoQ <- getQueenSegSiteGeno(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
genoF <- getFathersSegSiteGeno(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
genoW <- getWorkersSegSiteGeno(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
genoD <- getDronesSegSiteGeno(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
genoV <- getVirginQueensSegSiteGeno(apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
genoMeanW <- apply(X = genoW, MARGIN = 2, FUN = mean)
#> Error in apply(X = genoW, MARGIN = 2, FUN = mean): object 'genoW' not found
genoMeanD <- apply(X = genoD, MARGIN = 2, FUN = mean)
#> Error in apply(X = genoD, MARGIN = 2, FUN = mean): object 'genoD' not found

geno <- rbind(genoQ, genoF, genoW, genoD, genoV, genoMeanW, genoMeanD)
#> Error in rbind(genoQ, genoF, genoW, genoD, genoV, genoMeanW, genoMeanD): object 'genoQ' not found
n <- length(rownames(geno))
#> Error in rownames(geno): object 'geno' not found
rownames(geno)[c(n - 1, n)] <- c("mw", "md")
#> Error in rownames(geno)[c(n - 1, n)] <- c("mw", "md"): object 'geno' not found

sex <- getCasteSex(x = apiary[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
sex <- c(
  sex$queen, sex$fathers, sex$workers, sex$drones, sex$virginQueens,
  "F", "M"
)
#> Error in eval(expr, envir, enclos): object 'sex' not found

GRM <- calcBeeGRMIbs(x = geno, sex = sex)
#> Error in calcBeeGRMIbs(x = geno, sex = sex): object 'geno' not found

# library("Matrix"); image(as(GRM, "Matrix"))

x <- diag(GRM)
#> Error in diag(GRM): object 'GRM' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

x <- GRM[lower.tri(x = GRM, diag = FALSE)]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
hist(x)
#> Error in hist(x): object 'x' not found
summary(x)
#> Error in summary(x): object 'x' not found

q <- rownames(genoQ)
#> Error in rownames(genoQ): object 'genoQ' not found
f <- rownames(genoF)
#> Error in rownames(genoF): object 'genoF' not found
w <- rownames(genoW)
#> Error in rownames(genoW): object 'genoW' not found
d <- rownames(genoD)
#> Error in rownames(genoD): object 'genoD' not found
v <- rownames(genoV)
#> Error in rownames(genoV): object 'genoV' not found
mw <- "mw"
md <- "md"

# Queen vs others
GRM[q, f]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[q, w]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[q, d]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[q, v]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[q, mw]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[q, md]
#> Error in eval(expr, envir, enclos): object 'GRM' not found

# Fathers vs others
GRM[f, f]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[f, w]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[f, d]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[f, v]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[f, mw]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[f, md]
#> Error in eval(expr, envir, enclos): object 'GRM' not found

# Workers vs others
GRM[w, w]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[w, d]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[w, v]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[w, mw]
#> Error in eval(expr, envir, enclos): object 'GRM' not found
GRM[w, md]
#> Error in eval(expr, envir, enclos): object 'GRM' not found

# Calculating allele frequencies ourselves (say, to "shift" base population)
aF <- calcBeeAlleleFreq(x = geno, sex = sex)
#> Error in calcBeeAlleleFreq(x = geno, sex = sex): object 'geno' not found
hist(aF)
#> Error in hist(aF): object 'aF' not found
GRM2 <- calcBeeGRMIbs(x = geno, sex = sex, alleleFreq = aF)
#> Error in calcBeeGRMIbs(x = geno, sex = sex, alleleFreq = aF): object 'geno' not found
stopifnot(identical(GRM2, GRM))
#> Error in identical(GRM2, GRM): object 'GRM2' not found