Creates workers from the colony

Level 1 function that creates the specified number of workers from the colony by mating the colony queen and the fathers. If csd locus is active, it takes it into account and any csd homozygotes are removed and counted towards homozygous brood.

createWorkers(x, nInd = NULL, exact = FALSE, simParamBee = NULL)

Arguments

x

Colony-class or MultiColony-class

nInd

numeric or function, number of workers created; if NULL then SimParamBee$nWorkers is used

exact

logical, if the csd locus is active and exact is TRUE, create the exactly specified number of viable workers (heterozygous on the csd locus)

simParamBee

SimParamBee, global simulation parameters

Value

when x is Colony-class return is a list with two nodes named workers (a Pop-class) and nHomBrood

(a numeric); when x is MultiColony-class return is a list of lists named by colony ID

Examples

founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

drones <- createDrones(x = basePop[1], nInd = 1000)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
colony <- cross(colony, fathers = fatherGroups[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- createMultiColony(basePop[3:4], n = 2)
#> Error in createMultiColony(basePop[3:4], n = 2): object 'basePop' not found
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

(tmp <- createWorkers(colony, nInd = 10))
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
colony@queen@id
#> Error in eval(expr, envir, enclos): object 'colony' not found
tmp$workers@id
#> Error in eval(expr, envir, enclos): object 'tmp' not found
tmp$workers@sex
#> Error in eval(expr, envir, enclos): object 'tmp' not found
tmp$workers@misc
#> Error in eval(expr, envir, enclos): object 'tmp' not found
tmp$workers@mother
#> Error in eval(expr, envir, enclos): object 'tmp' not found
tmp$workers@father
#> Error in eval(expr, envir, enclos): object 'tmp' not found
SP$pedigree
#>      [,1] [,2] [,3]
head(SP$recHist)
#> list()
head(SP$recHist[[23]][[1]][1])
#> Error in SP$recHist[[23]]: subscript out of bounds
head(SP$recHist[[23]][[1]][2])
#> Error in SP$recHist[[23]]: subscript out of bounds
head(SP$caste)
#> factor(0)
#> Levels: queen fathers workers drones virginQueens

createWorkers(apiary, nInd = 10)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# Using a default in SP$nWorkers
# (just to have some workers - change this to your needs!)
createWorkers(colony)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
createWorkers(apiary)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

# Specify own number
SP$nWorkers <- 15
createWorkers(colony)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
createWorkers(apiary)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# nWorkers will NOT vary between function calls when a constant is used

# Specify a function that will give a number
createWorkers(colony, nInd = nWorkersPoisson)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
createWorkers(apiary, nInd = nWorkersPoisson)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# nWorkers will vary between function calls when a function is used

# Store a function or a value in the SP object
SP$nWorkers <- nWorkersPoisson
createWorkers(colony)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
createWorkers(apiary)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# nWorkers will vary between function calls when a function is used

# Inbred virgin queen with her brothers to generate csd homozygous brood
colony2 <- createColony(createVirginQueens(colony, nInd = 1))
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
colony2 <- cross(x = colony2, fathers = createDrones(colony, nInd = 15))
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# Check the expected csd homozygosity
pHomBrood(colony2)
#> Error in is(x, class2 = "Pop"): object 'colony2' not found

# Evaluate a realised csd homozygosity
createWorkers(colony2, nInd = 100)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# nHomBrood will vary between function calls due to inheritance process
createWorkers(colony2, nInd = 100)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# nHomBrood will vary between function calls due to inheritance process