Access epistasis values of individuals in a caste

Level 0 function that returns epistasis values of individuals in a caste.

getAa(x, caste = NULL, nInd = NULL, collapse = FALSE, simParamBee = NULL)

getQueenAa(x, collapse = FALSE, simParamBee = NULL)

getFathersAa(x, nInd = NULL, collapse = FALSE, simParamBee = NULL)

getVirginQueensAa(x, nInd = NULL, collapse = FALSE, simParamBee = NULL)

getWorkersAa(x, nInd = NULL, collapse = FALSE, simParamBee = NULL)

getDronesAa(x, nInd = NULL, collapse = FALSE, simParamBee = NULL)

Arguments

x

Pop-class, Colony-class, or MultiColony-class

caste

NULL or character, NULL when x is a Pop-class, and character when x is a Colony-class or MultiColony-class with the possible values of "queen", "fathers", "workers", "drones", "virginQueens", or "all"

nInd

numeric, number of individuals to access, if NULL all individuals are accessed, otherwise a random sample

collapse

logical, if the return value should be a single matrix with epistatic values of all the individuals

simParamBee

SimParamBee, global simulation parameters

Value

vector of epistasis values when x is

Colony-class and list of vectors of epistasis values when

x is MultiColony-class, named by colony id when x

is MultiColony-class

# Not exporting this function, since the theory behind it is not fully developed

Functions

getQueenAa(): Access epistasis value of the queen

getFathersAa(): Access epistasis values of fathers

getVirginQueensAa(): Access epistasis values of virgin queens

getWorkersAa(): Access epistasis values of workers

getDronesAa(): Access epistasis values of drones

See also