Set colony phenotype

Level 2 function that sets phenotypes for all colony individuals (queen, workers, drones, and virgin queens) and for the colony, or each colony in a MultiColony object.

setColonyPheno(x, FUN = NULL, reset = TRUE, simParamBee = NULL, ...)

Arguments

x

Colony-class or MultiColony-class

FUN

function, any function that can be run on colony and returns colony phenotypes; if NULL then SimParamBee$colonyPheno is used - if even this is NULL, then colony phenotype is not set, but phenotypes of colony individuals are (see reset though!)

reset

logical, should previous phenotype values for individuals in the castes be used or sampled anew?

simParamBee

SimParamBee, global simulation parameters

...

arguments passed to setPheno

Value

Colony-class or MultiColony-class with phenotypes

Details

When this function is called on a colony, phenotypes for all colony individuals and possibly the whole colony are set (or reset if phenotypes already exist).

See also

Examples

founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)

# Define two traits that collectively affect colony honey yield:
# 1) queen's effect on colony honey yield, say via pheromone secretion phenotype
# 2) workers' effect on colony honey yield, say via foraging phenotype
# The traits will have a negative genetic correlation of -0.5 and heritability
# of 0.25 (on an individual level)
meanP <- c(20, 0)
nWorkers <- 10
varA <- c(1, 1 / nWorkers)
corA <- matrix(data = c(
  1.0, -0.5,
  -0.5, 1.0
), nrow = 2, byrow = TRUE)
varE <- c(3, 3 / nWorkers)
varA / (varA + varE)
#> [1] 0.25 0.25
SP$addTraitA(nQtlPerChr = 100, mean = meanP, var = varA, corA = corA)

basePop <- createVirginQueens(founderGenomes)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
drones <- createDrones(x = basePop[1], nInd = 100)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = 14)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
colony <- cross(colony, fathers = fatherGroups[[1]])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
colony <- buildUp(colony, nWorkers = nWorkers, nDrones = 3)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- createMultiColony(basePop[3:5], n = 2)
#> Error in createMultiColony(basePop[3:5], n = 2): object 'basePop' not found
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
apiary <- buildUp(apiary, nWorkers = nWorkers, nDrones = 3)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

# Set phenotypes for all colony individuals
colony <- setColonyPheno(colony, varE = varE)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

apiary <- setColonyPheno(apiary, varE = varE)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found

# Queen's phenotype for both traits
getQueenPheno(colony)
#> Error in is(x, class2 = "Colony"): object 'colony' not found
getQueenPheno(apiary)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found

# Workers' phenotype for both traits
getWorkersPheno(colony)
#> Error in is(x, class2 = "Colony"): object 'colony' not found
getWorkersPheno(apiary)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found

# For the whole colony
getColonyPheno(colony)
#> Error in is(x, class2 = "Colony"): object 'colony' not found
getColonyPheno(apiary)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found

# Set phenotypes for all colony individuals AND Colony
colony <- setColonyPheno(colony, FUN = calcColonyPhenoFromCaste, varE = varE)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
getColonyPheno(colony)$colony
#> Error in is(x, class2 = "Colony"): object 'colony' not found

# Set phenotypes for all colony individuals AND MultiColony
apiary <- setColonyPheno(apiary, FUN = calcColonyPhenoFromCaste, varE = varE)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
sapply(X = getColonyPheno(apiary), FUN = function(x) x$colony)
#> Error in is(x, class2 = "Colony"): object 'apiary' not found

# Colony phenotype - store the colony function into the SP object
SP$colonyPheno <- calcColonyPhenoFromCaste
getColonyPheno(setColonyPheno(colony, varE = varE))$colony
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
getColonyPheno(setColonyPheno(colony, varE = varE))$colony
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
sapply(X = getColonyPheno(setColonyPheno(apiary, varE = varE)), FUN = function(x) x$colony)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
sapply(X = getColonyPheno(setColonyPheno(apiary, varE = varE)), FUN = function(x) x$colony)
#> Error in get(x = "SP", envir = .GlobalEnv): object 'SP' not found
# phenotype will vary between function calls by default (see reset)