Simulate the Honey bee genome

Level 0 function that returns simulated honeybee genomes

simulateHoneyBeeGenomes(
  nMelN = 0L,
  nMelS = 0L,
  nCar = 0L,
  nLig = 0L,
  ploidy = 2L,
  nChr = 16L,
  nSegSites = 100L,
  nBp = 225200000/16,
  genLen = 3.199121,
  mutRate = 3.4e-09,
  recRate = 2.3e-07,
  nThreads = NULL
)

Arguments

nMelN

integer, number of Apis mellifera mellifera North individuals to simulate

nMelS

integer, number of Apis mellifera mellifera South individuals to simulate

nCar

integer, number of Apis mellifera carnica individuals to simulate

nLig

integer, number of Apis mellifera ligustica individuals to simulate

ploidy

integer, the ploidy of the individuals

nChr

integer, number of chromosomes to simulate

nSegSites

integer, number of segregating sites to keep per chromosome

nBp

integer, base pair length of chromosome

genLen

numeric, genetic length of chromosome in Morgans

mutRate

numeric, per base pair mutation rate

recRate

numeric, per base pair recombination rate

nThreads

integer, if OpenMP is available, this will allow for simulating chromosomes in parallel. If NULL, the number of threads is automatically detected

Value

MapPop-class

References

Wallberg, A., Bunikis, I., Pettersson, O.V. et al. A hybrid de novo genome assembly of the honeybee, Apis mellifera, with chromosome-length scaffolds. 2019, BMC Genomics 20:275. https://doi.org/10.1186/s12864-019-5642-0

Beye M, Gattermeier I, Hasselmann M, et al. Exceptionally high levels of recombination across the honey bee genome. 2006, Genome Res 16(11):1339-1344. doi:10.1101/gr.5680406

Wallberg, A., Han, F., Wellhagen, G. et al. A worldwide survey of genome sequence variation provides insight into the evolutionary history of the honeybee Apis mellifera. 2014, Nat Genet 46:1081–1088. https://doi.org/10.1038/ng.3077

Yang S, Wang L, Huang J, Zhang X, Yuan Y, Chen JQ, Hurst LD, Tian D. Parent-progeny sequencing indicates higher mutation rates in heterozygotes. 2015, Nature 523(7561):463-7. doi: 10.1038/nature14649.

Examples

# founderGenomes <- simulateHoneyBeeGenomes(
#   nInd = 10, nChr = 1,
#   nSegSites = 2, Ne = 10
# )